refactor: reorganize serial communication classes and update imports
