+++ src/classes/__init__.py
@@ -1,6 +1,14 @@
-from debug import *
-from gamemode import *
-from input import *
-from output import *
-from parameter import *
-from serial import *
+if __name__ == '__main__':
+    from debug import *
+    from gameMode import *
+    from input import *
+    from output import *
+    from parameter import *
+    from serialHard import *
+else:
+    from classes.debug import *
+    from classes.gameMode import *
+    from classes.input import *
+    from classes.output import *
+    from classes.parameter import *
+    from classes.serialHard import *
